+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @whichly/web
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @whichly/react@0.1.1
+
 ## 0.0.1
 
 ### Patch Changes