+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @whichly/react
 
+## 0.1.1
+
+### Patch Changes
+
+- rename package
+
 ## 0.1.0
 
 ### Minor Changes